+++ cookbooks/ad-nativex/attributes/default.rb
@@ -1,14 +1,14 @@
 # Domain Name
-default['nativex-ad']['name'] = 'teamfreeze.com'
+default['ad-nativex']['name'] = 'teamfreeze.com'
 # OU
-default['nativex-ad']['ou'] = 'Domain Controlers'
+default['ad-nativex']['ou'] = 'Domain Controlers'
 # OUPath
-default['nativex-ad']['oupath'] = '"OU=Domain Controllers,DC=teamfreeze,DC=com"'
+default['ad-nativex']['oupath'] = '"OU=Domain Controllers,DC=teamfreeze,DC=com"'
 # Site Name
-default['nativex-ad']['site_name'] = 'AMAZON'
+default['ad-nativex']['site_name'] = 'AMAZON'
 # Safe Mode Password
-default['nativex-ad']['safe_mode_pass'] = 'Passw0rd'
+default['ad-nativex']['safe_mode_pass'] = 'Passw0rd'
 # AD User
-default['nativex-ad']['ad_username'] = 'nil'
+default['ad-nativex']['ad_username'] = 'nil'
 # AD Password
-default['nativex-ad']['ad_password'] = 'nil'
+default['ad-nativex']['ad_password'] = 'nil'