+++ cookbooks/ad-nativex/metadata.rb
@@ -1,11 +1,11 @@
-name             'nativex-ad'
+name             'ad-nativex'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures nativex-ad'
+description      'Installs/Configures ad-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 supports         'windows', ">= 6.2"
 
 depends		 "windows"
-depends		 "nativex-ohai"
+depends		 "ohai-nativex"