+++ cookbooks/ad-nativex/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-ad
+# Cookbook Name:: ad-nativex
 # Recipe:: default 
 #
 # Copyright 2014, NativeX
@@ -7,4 +7,4 @@
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe 'nativex-ad::joindomain'
+include_recipe 'ad-nativex::joindomain'