+++ cookbooks/ad-nativex/recipes/promotetodomaincontroller.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-ad
+# Cookbook Name:: ad-nativex
 # Recipe:: promotetodomaincontroller
 #
 # Copyright 2014, NativeX
@@ -11,12 +11,12 @@
 
 # Promote to Domain Controller
 ad = Chef::EncryptedDataBagItem.load("credentials", "ad")
-nativex_ad_domaincontroller "#{node['nativex-ad']['name']}" do
+nativex_ad_domaincontroller "#{node['ad-nativex']['name']}" do
   action :create
   type "replica" 
   domain_pass ad["ad_password"]
   domain_user ad["ad_username"]
-  site_name "#{node['nativex-ad']['site_name']}"
-  safe_mode_pass "#{node['nativex-ad']['safe_mode_pass']}"
+  site_name "#{node['ad-nativex']['site_name']}"
+  safe_mode_pass "#{node['ad-nativex']['safe_mode_pass']}"
 end